social_media_api/requirements.txt
asgiref==3.10.0
asn1crypto==1.5.1
base58==2.1.1
bip_utils==2.10.0
blinker==1.9.0
cbor2==5.7.1
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.1.8
coincurve==21.0.0
colorama==0.4.6
crcmod==1.7
dj-database-url==3.0.1
Django==5.2.7
django-filter==25.2
django-taggit==6.1.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
ecdsa==0.19.1
ed25519-blake2b==1.4.1
Flask==3.1.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mnemonic==0.21
mysql-connector-python==9.4.0
mysqlclient==2.2.7
pillow==12.0.0
protobuf==3.19.6
psycopg2-binary==2.9.11
py-sr25519-bindings==0.2.2
py_crypto_hd_wallet==1.3.3
pycparser==2.23
pycryptodome==3.19.1
PyJWT==2.10.1
PyNaCl==1.6.0
requests==2.32.5
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3
